Separate public from internal structs
