+++ lib/src/lib.rs
@@ -1,5 +1,8 @@
 mod model;
 mod persist;
+pub mod prelude;
+mod swap_in;
+mod swap_out;
 mod wallet;
 
 pub use model::*;
