+++ lib/src/prelude.rs
@@ -1,0 +1,4 @@
+//! Prelude, which exports all public structs and methods.
+
+pub use crate::model::*;
+pub use crate::wallet::*;
