+++ lib/src/swap_in/mod.rs
@@ -1,0 +1,1 @@
+//! Swap into LN from this wallet (L-BTC -> LN) using a submarine swap
