+++ lib/src/swap_out/error.rs
@@ -1,0 +1,41 @@
+use boltz_client::util::error::S5Error;
+
+#[derive(thiserror::Error, Debug)]
+pub enum SwapError {
+    #[error("Could not contact Boltz servers: {err}")]
+    ServersUnreachable { err: String },
+
+    #[error("Invoice amount is out of range")]
+    AmountOutOfRange,
+
+    #[error("Wrong response received from Boltz servers")]
+    BadResponse,
+
+    #[error("The specified invoice is not valid")]
+    InvalidInvoice,
+
+    #[error("Could not sign/send the transaction")]
+    SendError,
+
+    #[error("Could not fetch the required wallet information")]
+    WalletError,
+
+    #[error("Could not store the swap details locally")]
+    PersistError,
+
+    #[error("Generic boltz error: {err}")]
+    BoltzGeneric { err: String },
+}
+
+impl From<S5Error> for SwapError {
+    fn from(err: S5Error) -> Self {
+        match err.kind {
+            boltz_client::util::error::ErrorKind::Network
+            | boltz_client::util::error::ErrorKind::BoltzApi => {
+                SwapError::ServersUnreachable { err: err.message }
+            }
+            boltz_client::util::error::ErrorKind::Input => SwapError::BadResponse,
+            _ => SwapError::BoltzGeneric { err: err.message },
+        }
+    }
+}
