+++ lib/src/swap_out/mod.rs
@@ -1,0 +1,4 @@
+//! Swap out of LN and into this wallet (LN -> L-BTC) using a reverse swap
+
+pub(crate) mod error;
+pub(crate) mod model;
